Rename files and folders to new structure
