+++ App/Harp.Olfactometer.App/Harp.Olfactometer.App.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <OutputType>WinExe</OutputType>
@@ -41,7 +41,7 @@
   </Target>
 
   <ItemGroup>
-    <ProjectReference Include="..\Olfactometer.Design\Olfactometer.Design.csproj"/>
+    <ProjectReference Include="..\Harp.Olfactometer.Design\Harp.Olfactometer.Design.csproj"/>
   </ItemGroup>
 
   <ItemGroup>
