+++ App/Harp.Olfactometer.Design/App.axaml
@@ -1,0 +1,17 @@
+<Application xmlns="https://github.com/avaloniaui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:design="using:Olfactometer.Design"
+             x:Class="Olfactometer.Design.App"
+             Name="Olfactometer">
+
+    <Application.DataTemplates>
+        <design:ViewLocator />
+    </Application.DataTemplates>
+
+
+    <Application.Styles>
+        <FluentTheme Mode="Light" />
+        <StyleInclude Source="avares://Harp.Olfactometer.Design/Views/Icons.axaml" />
+        <StyleInclude Source="avares://Harp.Olfactometer.Design/Views/AccentColors.axaml" />
+    </Application.Styles>
+</Application>
