+++ App/Harp.Olfactometer.Design/Harp.Olfactometer.Design.csproj
@@ -1,0 +1,72 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <OutputType>Library</OutputType>
+    <TargetFramework>net6.0</TargetFramework>
+    <RuntimeIdentifiers>win-x64;osx-x64;linux-x64</RuntimeIdentifiers>
+    <RootNamespace>Olfactometer.Design</RootNamespace>
+    <Version>1.1.0</Version>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+  <PropertyGroup Condition=" '$(Configuration)' == 'Debug' ">
+    <OutputPath>../bin/Debug/</OutputPath>
+  </PropertyGroup>
+  <PropertyGroup Condition=" '$(Configuration)' == 'Release' ">
+    <OutputPath>../bin/Release/</OutputPath>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <Compile Update="**\*.xaml.cs">
+      <DependentUpon>%(Filename)</DependentUpon>
+    </Compile>
+    <AvaloniaResource Include="**\*.xaml">
+      <SubType>Designer</SubType>
+    </AvaloniaResource>
+    <AvaloniaResource Include="Assets\**"/>
+    <Compile Update="App.axaml.cs">
+      <DependentUpon>App.axaml</DependentUpon>
+      <SubType>Code</SubType>
+    </Compile>
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectCapability Include="Avalonia"/>
+    <TrimmerRootAssembly Include="Avalonia.Themes.Fluent"/>
+  </ItemGroup>
+
+  <ItemGroup>
+    <EmbeddedResource Include="Assets\Fonts\OpenSans-Bold.ttf"/>
+    <EmbeddedResource Include="Assets\Fonts\OpenSans-BoldItalic.ttf"/>
+    <EmbeddedResource Include="Assets\Fonts\OpenSans-ExtraBold.ttf"/>
+    <EmbeddedResource Include="Assets\Fonts\OpenSans-ExtraBoldItalic.ttf"/>
+    <EmbeddedResource Include="Assets\Fonts\OpenSans-Italic.ttf"/>
+    <EmbeddedResource Include="Assets\Fonts\OpenSans-Light.ttf"/>
+    <EmbeddedResource Include="Assets\Fonts\OpenSans-LightItalic.ttf"/>
+    <EmbeddedResource Include="Assets\Fonts\OpenSans-Regular.ttf"/>
+    <EmbeddedResource Include="Assets\Fonts\OpenSans-SemiBold.ttf"/>
+    <EmbeddedResource Include="Assets\Fonts\OpenSans-SemiBoldItalic.ttf"/>
+  </ItemGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Avalonia" Version="0.10.21"/>
+    <PackageReference Include="Avalonia.Desktop" Version="0.10.21"/>
+    <!--Condition below is needed to remove Avalonia.Diagnostics package from build output in Release configuration.-->
+    <PackageReference Condition="'$(Configuration)' == 'Debug'" Include="Avalonia.Diagnostics" Version="0.10.21"/>
+    <PackageReference Include="Avalonia.ReactiveUI" Version="0.10.21"/>
+    <PackageReference Include="Avalonia.ReactiveUI.Events" Version="0.10.21"/>
+    <PackageReference Include="Avalonia.Svg.Skia" Version="0.10.18"/>
+    <PackageReference Include="HexIO" Version="3.0.5"/>
+    <PackageReference Include="MessageBox.Avalonia" Version="2.1.0"/>
+    <PackageReference Include="ReactiveUI.Fody" Version="19.2.1"/>
+    <PackageReference Include="ReactiveUI.Validation" Version="3.1.7"/>
+    <PackageReference Include="System.IO.Ports" Version="6.0.0"/>
+    <PackageReference Include="XamlNameReferenceGenerator" Version="1.6.1"/>
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\Interface\Harp.Olfactometer\Harp.Olfactometer.csproj"/>
+  </ItemGroup>
+
+  <ItemGroup>
+    <Folder Include="Models\"/>
+  </ItemGroup>
+</Project>
