+++ App/Harp.Olfactometer.Design/Models/EepromManager.cs
