+++ App/Harp.Olfactometer.Design/MyCustomObservableExceptionHandler.cs
