+++ App/Harp.Olfactometer.Design/ViewModels/AboutViewModel.cs
