+++ App/Harp.Olfactometer.Design/ViewModels/EepromGenerationViewModel.cs
