+++ App/Harp.Olfactometer.Design/ViewModels/ViewModelBase.cs
