+++ App/Harp.Olfactometer.Design/Views/About.axaml.cs
