+++ App/Harp.Olfactometer.Design/Views/EepromGenerationWindow.axaml
