+++ App/Harp.Olfactometer.Design/Views/Icons.axaml
