+++ App/Harp.Olfactometer.Design/Views/MainWindow.axaml.cs
