+++ App/Harp.Olfactometer.Design/Views/MyDeviceView.axaml.cs
